+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+* Avoid mandating that the secrets directory exists.
+* Support configuration of the secrets directory using CLI arguments.
+
 ### [v1.10](https://github.com/realityforge/keycloak-converger/tree/v1.10) (2020-09-01) · [Full Changelog](https://github.com/realityforge/keycloak-converger/compare/v1.9...v1.10)
 
 * Save the client secrets as part of the converge process.